cxx: personalize the greeting function
As before, this introduces a bug, but at least in C++, it leads to a build
error

```
Undefined symbols for architecture x86_64:
  "__Z8greetingv", referenced from:
      _main in hello.o
ld: symbol(s) not found for architecture x86_64
collect2: error: ld returned 1 exit status
```

The reason being that in C++, functions can be overloaded, so it's legal
to have two versions of greeting() -- one which takes no arguments, and one
which takes the name. However, the former doesn't exist anymore, hence
the link error.